Make schedulers work with preload and small fixes
