+++ backend/src/controllers/reminders.py
@@ -1,5 +1,4 @@
 from flask import request
-from mail_scheduler import update_scheduler
 from utilities.sched_setting_methods import get_readable_settings, save_settings
 from utilities.require_login import require_login
 from utilities.require_admin import require_admin
@@ -16,7 +15,6 @@ def reminder_settings():
         try:
             data = request.json
             save_settings(data)
-            update_scheduler()
             return get_readable_settings()
         except ValueError as error:
             return str(error), 400
