+++ backend/src/mail_scheduler.py
@@ -1,34 +1,9 @@
-from datetime import date, timedelta, datetime
 from apscheduler.schedulers.background import BackgroundScheduler
 from apscheduler.triggers.cron import CronTrigger
-from flask_mail import Mail, Message
-from utilities import client_methods as clients
-from utilities.sched_setting_methods import load_settings, get_readable_settings, delete_custom
-from controllers.sms import send_sms_message
-from app import app
+from utilities.reminder_methods import send_email_reminders, send_sms_reminders
+from utilities.sched_setting_methods import load_settings, recover_settings
 
-mail = Mail(app)
 sched = BackgroundScheduler(daemon=True)
-sched.start()
-
-def send_email_reminders(remindertext):
-    with app.app_context():
-        deadlines, emails = get_reminder_data()[:2]
-        for deadline, email in zip(deadlines, emails): # pylint: disable=unused-variable
-            recipient = email
-            msg = Message('Muistutus lähestyvästä eräpäivästä',
-                        sender = app.config['MAIL_USERNAME'],
-                        recipients = [recipient])
-            msg.body = remindertext
-            mail.send(msg)
-
-def send_sms_reminders(remindertext):
-    with app.app_context():
-        _, _, phonenumbers = get_reminder_data()
-        for phonenumber in phonenumbers:
-            success = send_sms_message(phonenumber, remindertext, auto=True)
-            if not success:
-                print(f"Failed to send SMS to {phonenumber}")
 
 def update_scheduler(minute = 0, second = 0):
     settings = load_settings()
@@ -49,10 +24,6 @@ def update_scheduler(minute = 0, second = 0):
             settings = recover_settings()
 
     return False
-
-def recover_settings(filename = 'custom.json'):
-    delete_custom(filename)
-    return load_settings('default.json')
 
 def run_new_job(trigger, settings):
     existing_jobs = sched.get_jobs()
@@ -97,50 +68,8 @@ def create_trigger(
     )
     return trigger
 
-def get_reminder_data():
-    deadlines, client_ids = get_deadline_data()
-    emails = get_emails(client_ids)
-    phonenumbers = get_phonenumbers(client_ids)
-    return deadlines, emails, phonenumbers
-
-def get_deadline_data(client_service = clients):
-    settings = get_readable_settings()
-    deltas = [timedelta(days=delta) for delta in settings['deltas']]
-    deadlines_with_ids = client_service.get_next_deadlines()
-    deadlines = []
-    client_ids = []
-    today = datetime.today().weekday()
-    to_next_run = timedelta(days=days_to_next_run(settings['days'], today))
-    for deadline in deadlines_with_ids:
-        time_left = deadline.next_deadline - date.today()
-        if include_in_run(time_left, to_next_run, deltas):
-            deadlines.append(deadline.next_deadline)
-            client_ids.append(deadline.client_id)
-    return deadlines, client_ids
-
-def get_emails(client_ids):
-    emails = []
-    for c_id in client_ids:
-        emails.append(clients.get_email(c_id))
-    return emails
-
-def get_phonenumbers(client_ids):
-    numbers = []
-    for c_id in client_ids:
-        numbers.append(clients.get_phonenumber(c_id))
-    return numbers
-
 def list_jobs():
     return sched.print_jobs()
-def days_to_next_run(run_days, from_day):
-    for day in run_days:
-        if int(day) > int(from_day):
-            return int(day) - int(from_day)
 
-    return int(run_days[0]) + 7 - int(from_day)
-
-def include_in_run(time_left, to_next_run, deltas):
-    for delta in deltas:
-        if time_left - to_next_run < delta <= time_left:
-            return True
-    return False
+sched.add_job(update_scheduler, 'cron', minute=58)
+sched.start()
